# store-gateway: remove chunks ranges