--- a/pkg/storegateway/bucket_chunk_reader_test.go
+++ b/pkg/storegateway/bucket_chunk_reader_test.go
@@ -19,7 +19,7 @@ import (
 )
 
 func (g seriesChunkRefsRange) refAt(i int) chunks.ChunkRef {
-	return chunkRef(g.segmentFile, g.refs[i].segFileOffset)
+	return chunkRef(g.refs[i].segmentFile, g.refs[i].segFileOffset)
 }
 
 func TestBucketChunkReader_refetchChunks(t *testing.T) {