--- a/pkg/storegateway/bucket_chunk_reader_test.go
+++ b/pkg/storegateway/bucket_chunk_reader_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/go-kit/log"
 	"github.com/prometheus/prometheus/model/labels"
-	"github.com/prometheus/prometheus/tsdb/chunks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
@@ -17,10 +16,6 @@ import (
 	"github.com/grafana/mimir/pkg/util/pool"
 	"github.com/grafana/mimir/pkg/util/test"
 )
-
-func (g seriesChunkRefsRange) refAt(i int) chunks.ChunkRef {
-	return chunkRef(g.refs[i].segmentFile, g.refs[i].segFileOffset)
-}
 
 func TestBucketChunkReader_refetchChunks(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
@@ -71,16 +66,12 @@ func TestBucketChunkReader_refetchChunks(t *testing.T) {
 			chunkrModifiedLen := block.chunkReader(ctx)
 
 			for sIdx, seriesRef := range seriesRefs {
-				for _, r := range seriesRef.chunksRanges {
+				loadedChunksCorrectLen[sIdx].chks = append(loadedChunksCorrectLen[sIdx].chks, make([]storepb.AggrChunk, len(seriesRef.refs))...)
+				loadedChunksModifiedLen[sIdx].chks = append(loadedChunksModifiedLen[sIdx].chks, make([]storepb.AggrChunk, len(seriesRef.refs))...)
+				for _, r := range seriesRef.refs {
 					existingChunksNum := len(loadedChunksCorrectLen[sIdx].chks)
-
-					loadedChunksCorrectLen[sIdx].chks = append(loadedChunksCorrectLen[sIdx].chks, make([]storepb.AggrChunk, len(r.refs))...)
-					loadedChunksModifiedLen[sIdx].chks = append(loadedChunksModifiedLen[sIdx].chks, make([]storepb.AggrChunk, len(r.refs))...)
-
-					for cIdx, c := range r.refs {
-						assert.NoError(t, chunkrCorrectLen.addLoad(r.refAt(cIdx), sIdx, existingChunksNum+cIdx, c.length))
-						assert.NoError(t, chunkrModifiedLen.addLoad(r.refAt(cIdx), sIdx, existingChunksNum+cIdx, skewChunkLen(c.length)))
-					}
+					assert.NoError(t, chunkrCorrectLen.addLoad(r.storageRef(), sIdx, existingChunksNum, r.length))
+					assert.NoError(t, chunkrModifiedLen.addLoad(r.storageRef(), sIdx, existingChunksNum, skewChunkLen(r.length)))
 				}
 			}