… pkg/storegateway/series_refs_test.go