#### What this PR does

This is the next and final PR in removing fine-grained chunks caching from the store-gateway. Since we're no longer caching individual ranges of chunks we don't need to keep that abstraction in the code either. 

This removal also simplifies `loadingSeriesChunkRefsSetIterator`. The iterator no longer needs to partition chunks into ranges. It returns only chunks that fall within the requested minT/maxT.

In turn `loadingSeriesChunksSetIterator` no longer has to account for chunks that may fall outside of minT/maxT. I also took the liberty to consolidate the usage of `seriesIteratorStrategy` and update some outdated comments related to that.

This PR also adds back a test that was removed in https://github.com/grafana/mimir/pull/5877#issuecomment-1717677250

I didn't add a changelog entry because there shouldn't be any user- or operator-visible changes resulting from this PR.

#### Which issue(s) this PR fixes or relates to

Fixes #3939 

#### Checklist

- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`