@dimitarvdimitrov 2023-10-04
so the benchmarks on this branch vs main (71cfead16345544abc4f04a3cedecb800acc0c65) are very surprising; there are memory savings of up to 49% in some cases and CPU cahnges between -61% and +130%. I think this is due to this branch being very behind; i will rebase and run them again


<details><summary>Details</summary>
<p>

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/storegateway
cpu: Intel(R) Xeon(R) CPU @ 2.20GHz
                                                                                                  │  before.txt   │               after.txt                │
                                                                                                  │    sec/op     │     sec/op      vs base                │
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X"-4                                    21.02µ ±   1%    21.58µ ±   2%    +2.69% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j="foo"-4                            95.73µ ±  60%    39.68µ ±   2%         ~ (p=0.065 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j!="foo"-4                           39.23µ ± 146%    39.08µ ±   2%         ~ (p=0.485 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"XXX|YYY"-4                             5.931µ ±   1%    6.223µ ±   3%    +4.92% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"X.+"-4                                 5.807µ ±   1%    6.081µ ±   1%    +4.73% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"X|Y|Z"-4                               48.54µ ±   2%    49.56µ ±   2%    +2.09% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1"-4                                    408.2m ±   4%    375.3m ±   5%    -8.06% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j="foo"-4                            303.7m ±   2%    284.0m ±   1%    -6.49% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n="1"-4                            305.2m ±   3%    284.6m ±   3%    -6.77% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j!="foo"-4                           294.4m ±   3%    277.6m ±   3%    -5.69% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*"-4                                   7.975 ±  20%     7.492 ±  11%    -6.05% (p=0.041 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".+"-4                                  10.893 ±   7%     9.879 ±   4%    -9.31% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"^.+$",j=~"X.+"-4                       24.15m ±   6%    25.69m ±  22%    +6.41% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~""-4                                     2.595 ±   1%     2.646 ±   1%    +1.99% (p=0.009 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!=""-4                                    10.929 ±  11%     9.713 ±   5%   -11.13% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",j="foo"-4                    312.9m ±   4%    301.3m ±   5%    -3.73% (p=0.015 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",j="foo"-4                  39.45µ ±   1%    41.01µ ±   2%    +3.94% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",i!="2",j="foo"-4             331.0m ±   3%    310.4m ±   2%    -6.22% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!=""-4                               2.554 ±   2%     2.995 ±  17%         ~ (p=0.310 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j="foo"-4                       2.407 ±   2%     2.838 ±   2%   +17.88% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"X.+"-4                     23.38m ±   1%    25.79m ±   6%   +10.32% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"XXX|YYY"-4                 28.87µ ±   3%    29.81µ ±   2%    +3.26% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"X|Y|Z",j="foo"-4                 87.17µ ±   2%    87.92µ ±   1%         ~ (p=0.180 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",j="foo"-4                     2.427 ±   2%     2.436 ±   2%         ~ (p=0.394 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"1.+",j="foo"-4                   296.3m ±   3%    293.0m ±   3%         ~ (p=1.000 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!="2",j="foo"-4              2.440 ±   2%     2.449 ±   2%         ~ (p=0.699 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!~"2.*",j="foo"-4            2.658 ±   1%     2.654 ±  15%         ~ (p=0.589 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-4     39.50µ ±   3%    41.77µ ±   1%    +5.73% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"0xxx|1xxx|2xxx"-4                      621.0µ ±   3%    637.8µ ±   2%    +2.70% (p=0.026 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"(0|1|2)xxx"-4                          620.3µ ±   1%    639.5µ ±   2%    +3.09% (p=0.009 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"[0-2]xxx"-4                            621.1µ ±   4%    639.8µ ±   1%    +3.01% (p=0.041 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!~[0-2]xxx-4                               8.326 ±   8%     8.545 ±  93%         ~ (p=0.818 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*",_i!~[0-2]xxx-4                      8.012 ±   7%     7.085 ±   6%   -11.57% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"<unique_prefix>.+"-4                   262.1µ ±   2%    257.8µ ±   3%         ~ (p=0.132 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"<unique_prefix>.+"-4             6.180m ±  49%    3.202m ±   1%         ~ (p=0.180 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!~"<unique_prefix>.+"-4             995.5m ±  14%    379.0m ±   7%   -61.93% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",p="foo",i=~"<unique_prefix>.+"-4   59.26m ±   7%    30.94m ±   3%   -47.78% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n=~".+",i=~"<unique_prefix>.+"-4   172.4m ±  44%    103.1m ±   2%         ~ (p=0.394 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n=~".*",i=~"<unique_prefix>.+"-4   25.13m ±   5%    26.46m ±   3%    +5.29% (p=0.009 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~".*",n=~".*",i=~"<unique_prefix>.+"-4   310.8µ ±   6%    314.8µ ±   1%    +1.27% (p=0.026 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~".+",n=~".+",i=~"<unique_prefix>.+"-4   131.8m ±   2%    133.8m ±   3%         ~ (p=0.065 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/p!=""-4                                     1.727 ±  19%     1.555 ±  10%   -10.00% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X"-4                                       20.86µ ±   1%    22.32µ ±   3%    +6.97% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",j="foo"-4                               38.19µ ±   1%    40.51µ ±   1%    +6.09% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",j!="foo"-4                              38.05µ ±   3%    40.86µ ±   3%    +7.37% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~"XXX|YYY"-4                                5.802µ ±   1%    6.590µ ±   2%   +13.58% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~"X.+"-4                                    5.777µ ±  31%    6.470µ ±   1%         ~ (p=0.394 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"X|Y|Z"-4                                  48.69µ ±   2%    51.06µ ±   2%    +4.88% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1"-4                                       404.6m ±   3%    382.5m ±   5%    -5.45% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",j="foo"-4                               302.8m ±   3%    284.2m ±   2%    -6.14% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",n="1"-4                               298.7m ±   2%    284.0m ±   2%    -4.94% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",j!="foo"-4                              294.9m ±   2%    276.6m ±   4%    -6.19% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".*"-4                                      8.229 ±  21%     7.279 ±   7%   -11.54% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".+"-4                                     10.453 ±   5%     9.652 ±   5%    -7.66% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"^.+$",j=~"X.+"-4                          23.23m ±   6%    25.13m ±   8%         ~ (p=0.065 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~""-4                                        2.593 ±   2%     2.598 ±   2%         ~ (p=0.937 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i!=""-4                                       10.831 ±   6%     9.815 ±  16%         ~ (p=0.065 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".*",j="foo"-4                       315.4m ±   5%    298.4m ±   5%    -5.39% (p=0.041 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",i=~"^.+$",j="foo"-4                     39.88µ ±   2%    41.11µ ±   4%    +3.09% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".*",i!="2",j="foo"-4                328.5m ±   2%    306.1m ±   1%    -6.82% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!=""-4                                  2.534 ±   2%     2.501 ±   2%         ~ (p=0.093 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j="foo"-4                          2.413 ±   2%     2.441 ±   1%    +1.16% (p=0.026 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j=~"X.+"-4                        23.04m ±   2%    24.34m ±   1%    +5.65% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!="",j=~"XXX|YYY"-4                    29.88µ ±   3%    30.45µ ±   1%         ~ (p=0.132 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"X|Y|Z",j="foo"-4                    88.53µ ±   2%    88.95µ ±   1%         ~ (p=0.699 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",j="foo"-4                        2.442 ±   3%     2.413 ±   1%         ~ (p=0.394 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"1.+",j="foo"-4                      293.1m ±   2%    292.4m ±   1%         ~ (p=0.699 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",i!="2",j="foo"-4                 2.434 ±   1%     2.438 ±   1%         ~ (p=0.699 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~".+",i!~"2.*",j="foo"-4               2.626 ±   2%     2.664 ±   1%    +1.45% (p=0.041 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-4        39.68µ ±   1%    40.78µ ±   1%    +2.76% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"0xxx|1xxx|2xxx"-4                         615.8µ ±   2%    730.0µ ± 129%         ~ (p=0.132 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"(0|1|2)xxx"-4                             615.3µ ±   4%   1417.3µ ±  16%  +130.33% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"[0-2]xxx"-4                               624.3µ ±   3%   1345.3µ ±  17%  +115.50% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i!~[0-2]xxx-4                                  7.937 ±  11%     7.181 ± 195%         ~ (p=0.394 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~".*",_i!~[0-2]xxx-4                        11.109 ±  67%     7.102 ±   7%   -36.07% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/i=~"<unique_prefix>.+"-4                      272.2µ ±   4%    260.5µ ±   5%    -4.28% (p=0.009 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i=~"<unique_prefix>.+"-4                3.196m ±   2%    3.229m ±   1%    +1.01% (p=0.041 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/n="1",i!~"<unique_prefix>.+"-4                402.7m ±   4%    375.1m ±   5%    -6.83% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",p="foo",i=~"<unique_prefix>.+"-4      30.16m ±   4%    30.93m ±   3%         ~ (p=0.240 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",n=~".+",i=~"<unique_prefix>.+"-4      97.06m ±   2%    99.06m ±   1%    +2.06% (p=0.026 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j="foo",n=~".*",i=~"<unique_prefix>.+"-4      25.96m ±   2%    26.23m ±   3%         ~ (p=0.394 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~".*",n=~".*",i=~"<unique_prefix>.+"-4      319.2µ ±   3%    326.3µ ±   7%         ~ (p=0.065 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/j=~".+",n=~".+",i=~"<unique_prefix>.+"-4      134.0m ±   3%    131.6m ±   2%         ~ (p=0.065 n=6)
OpenBlockSeriesChunkRefsSetsIterator/with_index_cache/p!=""-4                                        1.742 ±   6%     1.594 ±  12%    -8.51% (p=0.015 n=6)
geomean                                                                                             23.13m           22.45m           -2.91%

                                                                                                  │  before.txt  │              after.txt               │
                                                                                                  │     B/op     │     B/op      vs base                │
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X"-4                                    2.276Ki ± 0%   2.276Ki ± 0%       ~ (p=1.000 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j="foo"-4                            2.933Ki ± 0%   2.933Ki ± 0%       ~ (p=1.000 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",j!="foo"-4                           2.933Ki ± 0%   2.933Ki ± 0%       ~ (p=0.697 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"XXX|YYY"-4                             1.945Ki ± 0%   1.945Ki ± 0%       ~ (p=1.000 n=6) ¹
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j=~"X.+"-4                                 1.891Ki ± 0%   1.891Ki ± 0%       ~ (p=1.000 n=6) ¹
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"X|Y|Z"-4                               2.869Ki ± 0%   2.869Ki ± 0%       ~ (p=1.000 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1"-4                                    145.4Mi ± 0%   136.2Mi ± 0%  -6.32% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j="foo"-4                            92.52Mi ± 1%   88.03Mi ± 0%  -4.86% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",n="1"-4                            92.64Mi ± 0%   88.00Mi ± 1%  -5.01% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",j!="foo"-4                           92.72Mi ± 0%   88.06Mi ± 0%  -5.02% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*"-4                                  3.286Gi ± 0%   3.063Gi ± 0%  -6.80% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".+"-4                                  3.436Gi ± 0%   3.211Gi ± 0%  -6.52% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"^.+$",j=~"X.+"-4                       10.69Mi ± 0%   10.69Mi ± 0%  +0.00% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~""-4                                    224.7Mi ± 0%   224.7Mi ± 0%       ~ (p=0.180 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!=""-4                                    3.435Gi ± 0%   3.212Gi ± 0%  -6.50% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",j="foo"-4                    95.76Mi ± 0%   91.24Mi ± 1%  -4.72% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",j="foo"-4                  3.104Ki ± 0%   3.105Ki ± 0%       ~ (p=0.545 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".*",i!="2",j="foo"-4             96.97Mi ± 0%   92.51Mi ± 0%  -4.60% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!=""-4                              333.1Mi ± 0%   323.6Mi ± 0%  -2.87% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j="foo"-4                      279.7Mi ± 0%   275.0Mi ± 0%  -1.67% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"X.+"-4                     10.69Mi ± 0%   10.69Mi ± 0%       ~ (p=0.818 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!="",j=~"XXX|YYY"-4                 3.871Ki ± 0%   3.871Ki ± 0%       ~ (p=1.000 n=6) ¹
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"X|Y|Z",j="foo"-4                 5.123Ki ± 0%   5.122Ki ± 0%       ~ (p=0.764 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",j="foo"-4                    279.7Mi ± 0%   275.2Mi ± 0%  -1.58% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"1.+",j="foo"-4                   43.30Mi ± 0%   42.73Mi ± 0%  -1.32% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!="2",j="foo"-4             281.3Mi ± 0%   276.6Mi ± 0%  -1.68% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~".+",i!~"2.*",j="foo"-4           287.0Mi ± 0%   282.9Mi ± 0%  -1.44% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-4     3.308Ki ± 0%   3.308Ki ± 0%       ~ (p=1.000 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"0xxx|1xxx|2xxx"-4                      151.3Ki ± 0%   144.2Ki ± 0%  -4.72% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"(0|1|2)xxx"-4                          151.1Ki ± 0%   143.8Ki ± 0%  -4.81% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"[0-2]xxx"-4                            151.1Ki ± 0%   143.9Ki ± 0%  -4.71% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i!~[0-2]xxx-4                              3.283Gi ± 0%   3.059Gi ± 0%  -6.81% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~".*",_i!~[0-2]xxx-4                     3.285Gi ± 0%   3.062Gi ± 0%  -6.80% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/i=~"<unique_prefix>.+"-4                   56.40Ki ± 0%   54.08Ki ± 0%  -4.10% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i=~"<unique_prefix>.+"-4             1.273Mi ± 0%   1.273Mi ± 0%       ~ (p=0.818 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/n="1",i!~"<unique_prefix>.+"-4             145.7Mi ± 0%   136.2Mi ± 0%  -6.58% (p=0.002 n=6)
OpenBlockSeriesChunkRefsSetsIterator/without_index_cache/j="foo",p="foo",i=~"<unique_prefix>.+"-4   18.45Mi ± 0%   18.46Mi ± 0%       ~ (p=0.132 n=6)
…